APP-596: [Extension] [send] add some icons in header for send screen.
